+++ frontend/src/App.jsx
@@ -1,22 +1,22 @@
 import { Route, BrowserRouter as Router, Routes } from "react-router-dom";
 import { useSelector } from "react-redux";
-import NavBar from "./NavBar";
-import LandingPage from "./LandingPage";
-import LoginPage from "./LoginPage";
-import SignupPage from "./SignupPage";
-import HomePage from "./HomePage";
-import PasswordGenPage from "./PasswordGenPage";
-import EncryptionTestPage from "./EncryptionTestPage";
-import NewPasswordPage from "./NewPasswordPage";
-import NewCreditCardPage from "./NewCreditCardPage";
-import NewNotePage from "./NewNotePage";
-import CacheTestPage from "./cacheTestPage";
-import DataExport from "./DataExport";
-import DataImport from "./DataImport";
-import EditPasswordPage from "./EditPasswordPage";
-import EditNotePage from "./EditNotePage";
-import EditCreditCardPage from "./EditCreditCardPage";
-import TrashBin from "./TrashBin";
+import NavBar from "./components/NavBar";
+import LandingPage from "./pages/LandingPage";
+import LoginPage from "./pages/LoginPage";
+import SignupPage from "./pages/SignupPage";
+import HomePage from "./pages/HomePage";
+import PasswordGenPage from "./pages/PasswordGenPage";
+import EncryptionTestPage from "./pages/EncryptionTestPage";
+import NewPasswordPage from "./pages/NewPasswordPage";
+import NewCreditCardPage from "./pages/NewCreditCardPage";
+import NewNotePage from "./pages/NewNotePage";
+import CacheTestPage from "./pages/cacheTestPage";
+import DataExport from "./pages/DataExportPage";
+import DataImport from "./pages/DataImportPage";
+import EditPasswordPage from "./pages/EditPasswordPage";
+import EditNotePage from "./pages/EditNotePage";
+import EditCreditCardPage from "./pages/EditCreditCardPage";
+import TrashBin from "./pages/TrashBinPage";
 
 function App() {
   const loggedInUser = useSelector((state) => state.auth.user);
