+++ frontend/src/components/DeleteConfirmation.jsx
