+++ frontend/src/components/DeleteConfirmationTrash.jsx
