+++ frontend/src/components/NavBar.jsx
@@ -1,11 +1,18 @@
-import React from 'react';
-import { Box, Typography, ButtonBase, Button, Menu, MenuItem } from "@mui/material";
-import DeleteAccountConfirmationDialog from './DeleteAccountConfirmation';
+import React from "react";
+import {
+  Box,
+  Typography,
+  ButtonBase,
+  Button,
+  Menu,
+  MenuItem,
+} from "@mui/material";
+import DeleteAccountConfirmationDialog from "./DeleteAccountConfirmation";
 import { useNavigate } from "react-router-dom";
 import { useSelector, useDispatch } from "react-redux";
-import { logout } from "./slices/authSlice";
-import { clearInfo } from "./slices/userInfoSlice";
-import BeaverLogo from "./imgs/beaver_logo.png";
+import { logout } from "../slices/authSlice";
+import { clearInfo } from "../slices/userInfoSlice";
+import BeaverLogo from "../imgs/beaver_logo.png";
 
 export default function NavBar() {
   const navigate = useNavigate();
@@ -13,7 +20,7 @@ export default function NavBar() {
 
   const loggedInUser = useSelector((state) => state.auth.user);
   const accessToken = useSelector((state) => state.auth.accessToken);
-  
+
   const [deleteDialogOpen, setDeleteDialogOpen] = React.useState(false);
   const [anchorEl, setAnchorEl] = React.useState(null);
   const buttonRef = React.useRef(null);
@@ -40,26 +47,33 @@ export default function NavBar() {
         justifyContent: "space-between",
         gap: "1rem",
         padding: "1rem",
-        position: 'relative',
+        position: "relative",
       }}
     >
       <ButtonBase onClick={() => navigate("/")}>
         <img src={BeaverLogo} alt="Beaver Logo" width={100} />
-        <Typography variant="h5" sx={{ fontFamily: 'Roboto', fontWeight: 'bold' }}>Beaver Vault</Typography>     
+        <Typography
+          variant="h5"
+          sx={{ fontFamily: "Roboto", fontWeight: "bold" }}
+        >
+          Beaver Vault
+        </Typography>
       </ButtonBase>
       {accessToken && (
         <Box
           sx={{
-            position: 'absolute',
+            position: "absolute",
             right: 0,
             top: 0,
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'flex-end',
-            padding: '1.0rem',
+            display: "flex",
+            flexDirection: "column",
+            alignItems: "flex-end",
+            padding: "1.0rem",
           }}
         >
-          <Typography variant="h6" sx={{ padding: '0.5rem' }}>{loggedInUser["email"]}</Typography>
+          <Typography variant="h6" sx={{ padding: "0.5rem" }}>
+            {loggedInUser["email"]}
+          </Typography>
           <Button variant="contained" onClick={handleClick} ref={buttonRef}>
             Settings
           </Button>
@@ -69,102 +83,122 @@ export default function NavBar() {
             onClose={handleClose}
             sx={{ mt: 1 }}
           >
-            <MenuItem 
-              onClick={() => { 
-                navigate('/');
+            <MenuItem
+              onClick={() => {
+                navigate("/");
                 handleClose();
               }}
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Homepage
             </MenuItem>
 
-            <MenuItem 
-              onClick={() => { 
-                navigate('/');
+            <MenuItem
+              onClick={() => {
+                navigate("/");
                 handleClose();
-              }} 
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Change Email
             </MenuItem>
 
-            <MenuItem 
-              onClick={() => { 
-                navigate('/');
+            <MenuItem
+              onClick={() => {
+                navigate("/");
                 handleClose();
-              }} 
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Change Master Password
             </MenuItem>
 
-            <MenuItem 
-              onClick={() => { 
-                navigate('/cache-test');
+            <MenuItem
+              onClick={() => {
+                navigate("/cache-test");
                 handleClose();
-              }} 
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Cache Testing
             </MenuItem>
 
-            <MenuItem 
+            <MenuItem
               onClick={() => {
-                navigate('/passwordgen');
+                navigate("/passwordgen");
                 handleClose();
-              }} 
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Generate Password
             </MenuItem>
 
-            <MenuItem 
-              onClick={() => { 
+            <MenuItem
+              onClick={() => {
                 setDeleteDialogOpen(true);
                 handleClose();
-              }} 
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Delete Account
             </MenuItem>
 
-            <MenuItem 
-              onClick={() => { 
-                navigate('/dataimport');
+            <MenuItem
+              onClick={() => {
+                navigate("/dataimport");
                 handleClose();
               }}
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Import Vault
             </MenuItem>
 
-            <MenuItem 
-              onClick={() => { 
-                navigate('/dataexport');
+            <MenuItem
+              onClick={() => {
+                navigate("/dataexport");
                 handleClose();
               }}
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Export Vault
             </MenuItem>
 
-            <MenuItem 
-              onClick={() => { 
-                navigate('/trashbin');
+            <MenuItem
+              onClick={() => {
+                navigate("/trashbin");
                 handleClose();
               }}
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Trash Bin
             </MenuItem>
-                        
-            <MenuItem 
-              onClick={() => { 
+
+            <MenuItem
+              onClick={() => {
                 handleSignOut();
                 handleClose();
-              }} 
-              sx={{ minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0 }}
+              }}
+              sx={{
+                minWidth: buttonRef.current ? buttonRef.current.offsetWidth : 0,
+              }}
             >
               Sign Out
             </MenuItem>
