+++ frontend/src/components/PasswordCell.jsx
