+++ frontend/src/pages/EditNotePage.jsx
@@ -1,3 +1,4 @@
+import React, { useState, useEffect } from "react";
 import {
   Box,
   Typography,
@@ -6,38 +7,70 @@ import {
   TextField,
   Button,
 } from "@mui/material";
-import { useState } from "react";
 import { useSelector } from "react-redux";
-import { encryptText } from "./encryption";
-import { useNavigate } from "react-router-dom";
-import { useAddNoteMutation } from "./slices/apiSlice";
+import { encryptText, decryptText } from "../scripts/encryption";
+import { useNavigate, useParams } from "react-router-dom";
+import { useGetNotesQuery, useUpdateNoteMutation } from "../slices/apiSlice";
 
-export default function NewNotePage() {
+export default function EditNotePage() {
   const navigate = useNavigate();
-
+  const { id } = useParams();
   const loggedInUser = useSelector((state) => state.auth.user);
   const userFolders = useSelector((state) => state.userInfo.folders);
-  const accessToken = useSelector((state) => state.auth.accessToken);
 
-  const [addNotePost, noteResult] = useAddNoteMutation();
-
-  const [currentFolder, setCurrentFolder] = useState(userFolders[0].folderID);
+  const [currentFolder, setCurrentFolder] = useState(
+    userFolders.length > 0 ? userFolders[0].folderID : ""
+  );
   const [notename, setNotename] = useState("");
   const [content, setContent] = useState("");
 
+  const { data: noteData, refetch: noteRefetch } =
+    useGetNotesQuery(currentFolder);
+  const [updateNote] = useUpdateNoteMutation();
+
+  const fetchNoteData = async () => {
+    noteRefetch();
+    try {
+      if (!loggedInUser) {
+        console.error("User not logged in");
+        navigate("/");
+        return;
+      }
+      if (!noteData) {
+        return;
+      }
+
+      const matchedNote = noteData.find((note) => note.noteID === parseInt(id));
+      if (!matchedNote) {
+        console.error("Note not found");
+        navigate("/");
+        return;
+      }
+      setCurrentFolder(matchedNote.folderID);
+      setNotename(decryptText(matchedNote.noteName, loggedInUser.masterKey));
+      setContent(decryptText(matchedNote.content, loggedInUser.masterKey));
+    } catch (error) {
+      console.error("Error fetching note data:", error);
+    }
+  };
+
+  useEffect(() => {
+    fetchNoteData();
+  }, [id, loggedInUser.folderID, noteData]);
+
   const handleSubmit = async () => {
-    const noteData = {
+    const updatedNoteData = {
       folderID: currentFolder,
       noteName: encryptText(notename, loggedInUser.masterKey),
       content: encryptText(content, loggedInUser.masterKey),
     };
 
     try {
-      await addNotePost(noteData);
-      alert("Note added successfully");
+      await updateNote({ notesID: id, updatedData: updatedNoteData });
+      alert("Note updated successfully");
       navigate("/");
     } catch (error) {
-      console.error("Error adding note:", error);
+      console.error("Error updating note:", error);
       alert("An unexpected error occurred. Please try again.");
     }
   };
@@ -54,7 +87,7 @@ export default function NewNotePage() {
         margin: "auto",
       }}
     >
-      <Typography variant="h4">Add New Note</Typography>
+      <Typography variant="h4">Edit Note</Typography>
       <Select
         fullWidth
         value={currentFolder}
@@ -74,6 +107,7 @@ export default function NewNotePage() {
         fullWidth
         variant="outlined"
         label="Note Name"
+        value={notename}
         onChange={(e) => {
           setNotename(e.target.value);
         }}
@@ -82,6 +116,7 @@ export default function NewNotePage() {
         fullWidth
         variant="outlined"
         label="Content"
+        value={content}
         onChange={(e) => {
           setContent(e.target.value);
         }}
