+++ frontend/src/scripts/encryption.js
