+++ frontend/src/scripts/storage.js
