+++ frontend/src/scripts/submitPasswordToAPI.js
